requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyJWT
beautifulsoup4